requirements.txt
httpx
psycopg2-binary
fastapi
uvicorn
python-dotenv
asyncpg
